+++ src/main/java/module-info.java
@@ -1,4 +1,6 @@
 open module me.cdh {
     requires kotlin.stdlib;
     requires java.desktop;
+    requires kotlinx.coroutines.core;
+    requires kotlinx.coroutines.swing;
 }
