--- src/main/kotlin/Animate.kt
@@ -1,6 +1,0 @@
-package me.cdh
-
-interface Animate {
-    val delay:Int
-    val frame: Int
-}
