+++ src/main/kotlin/Constants.kt
@@ -1,0 +1,9 @@
+package me.cdh
+
+
+const val WINDOW_WIDTH = 100
+const val WINDOW_HEIGHT = 100
+const val ANIMATION_UPDATE_DELAY = 10L
+const val DAYTIME_WANDER_INTERVAL = 6000L
+const val NIGHTTIME_WANDER_INTERVAL = 30000L
+const val BUBBLE_SIZE = 30
