--- src/main/kotlin/enumerate/BubbleState.kt
@@ -1,12 +1,0 @@
-package me.cdh.enumerate
-
-import me.cdh.Animate
-
-enum class BubbleState(
-    override val frame: Int,
-    override val delay: Int
-) : Animate {
-    ZZZ(4, 30),
-    HEART(4, 50),
-    NONE(-1, -1);
-}
