--- src/main/kotlin/enumerate/Direction.kt
@@ -1,6 +1,0 @@
-package me.cdh.enumerate
-
-enum class Direction {
-    RIGHT,
-    LEFT
-}
