--- src/main/kotlin/enumerate/State.kt
@@ -1,6 +1,0 @@
-package me.cdh.enumerate
-
-enum class State {
-    DEFAULT,
-    WANDER
-}
